move tab extension points to tab app
